move encryption functions to new file
